refactor: remove games data to a separate module for avoiding the code duplication
