+++ rslang/src/data/games-data.ts
@@ -1,0 +1,17 @@
+import sprintImage from './../assets/images/sprint.png';
+import audioCallImage from './../assets/images/audio-call.png';
+
+const gamesData = [
+  { 
+    name: 'Спринт', 
+    description: 'Спринт - тренировка на скорость, В Вашем распоряжении 30 секунд, за которые необходимо угадать как можно больше слов', 
+    image: sprintImage 
+  },
+  { 
+    name: 'Аудио-вызов', 
+    description: 'Аудио-вызов - тренировка на слух: улучшает восприятие слов, используя для этого такой метод обучения как аудирование', 
+    image: audioCallImage
+  }
+];
+
+export default gamesData;
